moved test_breach_checker.py -> test_utils.py
